Portal history JSON-RPC: Add validation, where reasonable
Add content validation + local storing to Portal history JSON-RPC
methods:
- portal_historyGetContent
- portal_historyTraceGetContent

This means that for bodies and receipts being requested an
additional network lookup (the header) is done. This is likely
acceptable as they are considered high level calls.

Not adding the validation however for portal_historyPutContent
method as this would highly limit the bridge gossip efficiency.

In general the Portal JSON-RPC API specifications are not very
clear on which type of validation checks should be done in these
methods.