Tidying up wire protocol handler and add limit (#3116)
Unfortunately, because of rlp encoding, the rate limiter like
in nimbus-eth2 cannot be applied to nimbus-eth1 wire protocol
handler immediately. So we stick with what the spec says.

In nimbus-eth2 or consensus layer in general, the wire protocol is
encoded in SSZ. So it is possible to send the message in chunk relatively easy.

While execution client is using RLP encoding. Even before sending the
wire message, the full payload length prefix must be calculated,
therefore a special RLP encoder is needed. And because the full payload
must be held in memory, benefits of incremental read from database
cannot be applied.