Clean `lastSnapshots` when finalizing
When cleaning snapshots, it may happen that frames have already been
written (via base updates) leading to a crash when trying to clean up
their snapshot - `deques` doesn't readily support removing items in the
middle so use a circular buffer instead.

Also fixes an issue where a partial database write may lead to a
"future" block header may turn up in the database - when this happens,
we don't have an in-memory state that can be advanced and the next block
will fail to apply. Most likely, this happens because of how the syncer
shortcuts txframe updates when writing headers, which in and of itself
is a problem that needs a proper solution - in the meantime however,
checking that the state exists allows recovering from such an
inconsistent database.