Remove LeafTie, PathID (#3178)
* Remove LeafTie, PathID

These are mostly unused, except for implementing a complex version of
trie iteration - replacing the iteration with a simpler and more
performant version which operates on nibbles directly (instead of going
via an integer), we can remove this alternate represenation.

Along with it, remove obsolete testing code as well that has grown stale
following other database changes.

* comment

* oops