`aristo_part` -> `aristo_proof`, rm "partial database" support (#3181)
`aristo_part` seems to have been an incomplete attempt at storing
incomplete state databases as happens for example during snap sync but
also when you have a partial view of the state in general and fill in
the rest with proof nodes.

Apart from not working and not being tested, the current code is
obsolete and designed around "generic" storage nodes which no longer
exist - a fresh attempt will have to be made when snap sync is done - in
the meantime, we can extract the useful parts (proof generation) into
its own, hopefully better named, module.

When reimplemnting this support, _probably_ what needs to be done is to
introduce a new VertexRef "proof" type that only has a merkle key and no
data - even with partial databases, we're still operating on the
assumption that either accounts or slots are being stored and nothing
else - should there ever be need to store anything else but these two
data types, we should not use aristo at all.