Fix bad merge of `seq[byte]` resulting in `gasUsed mismatch` during s… (#3189)
This PR fixes an annoying and spurious issue that would occur when
merging seq[byte]-based kvt entries - EVM `code` in particular - when
syncing due to the accidental use of a "merge" function that joins the
two seqs instead of overwriting them.

Under the condition that the same code gets updated in _separate_ blocks
but within a single database checkpoint (such as happens when syncing
and updating base every 32 blocks), this would lead to an invalid code
snippet being written to the database. When the entry expired for the
code cache and got loaded again, it would lead to a `gasUsed mismatch`
in some random block down the line that was unlucky enough to load the
bad code from the database.