evm: less bounds checks in hotspot
`invalidPosition` is a hotspot (2% cpu when processing blocks) - this
tiny fix improves the bounds checking further by extending its avoidance
to the invalid position list whose length is determined by the code
length and therefore safe to access in valid code positions.

Doesn't make a huge difference but it's also not nothing - might as well
do it properly.