Switch VertexRef to inheritance (#3197)
The current VertexRef is a composition of two case object types - this
layout is inefficient in that each VertexRef instance takes up as much
memory as its largest member. Adding to the insult are the two case
fields which each take up space.

Here, we use inhertance instead to create optimally sized verticies for
each node kind - in particular, the Brach kind kan drop the prefix and
reduce its memory usage to a fraction of what it was using before.

The memory savings are instead invested in a larger key cache - when
processing blocks for live syncing, key lookups are the source of a
significant majority (95%) of database reads - when importing, the key
cache is simply allotted to verticies instead.