Delay state root checks for known-finalized blocks (fixes #3096) (#3183)
* Delay state root checks for known-finalized blocks

When a majority of validators (according to forkChoiceUpdated) have
verified a block, it's reasonable to skip or at least delay some
verification so that it can be done in bulk / faster.

Here, we delay the state root verification, doing it when it's time to
update base instead of eagerly when processing the block - with a
32-block window, this leads to a speedup of live syncing of around 2-3x,
with potential for more with a larger window.

We run the state root computation on a regular basis not only as a
sanity check but also to populate the database and not end up with "too
much" work at the end - there's lots of follow-up tuning that can be
done here.

* Resolve merge conflict

* Update test_forked_chain

* Verify state root in batch of 4 blocks

---------

Co-authored-by: jangko <jangko128@gmail.com>