Remove account leaf cache
With static vid lookup we can find cached accounts in `O(1)` in the
regular caches - no need to waste time and effort on maintaining a
separate path-based cache for it.

Without the account leaf cache, the number of queries for non-existing
vids spikes because of all the path-based guesses, which we compensate
with a simpler "empty" cache in the database layer.

We put the memory savings towards the branch/vtx caches instead.