Add access from History network to historical summaries
Add access from History network to historical summaries for the
verification of Capella and onwards block proofs.

Access is provided by adding the BeaconDbCache to the history
network, more specifical to the HeaderVerifier (before called
Accumulators). This approach is taken, over providing callbacks,
 as it is more in sync with how StateNetwork accesses the
HistoryNetwork. It might be still be considered to move to
callbacks in the future though as that could provide a more
"oracle" agnostic way of providing this data.

The BeaconDbCache is created because for Ephemeral headers
verification we will also need access to the Light client updates.
Aside from the Light client updates, the historical summaries
are also added to the cache in its decoded form for easy and
fast access on block verification.

Some changes are likely to be still required to avoid to many
copies of the summaries, TBI.