rm useless windeps.dll downloading; make syncer batch size --debug (#3328)
* rm useless windeps.dll downloading; make syncer batch size --debug

* add arm64 platform detection



The reasons for removing windeps.zip downloading are primarily twofold:

    it's a useless/unused network fetch which could fail; and
    given that, that it contains an EOL OpenSSL 1.1 DLL makes it potential attack surface to leave in the CI path.

There's no point (at least initially for sure, but probably ever) in supporting x86 macOS. It's rapidly fading. If nothing else, the defaults should be switched to ARM macOS, and if x86 really needs to be re-added, that can be looked at later. But right now the CI should be focusing on testing ARM.

--persist-batch-size looks very much like a debugging option we probably won't want to ship/document as an end-user supported command-line flag, so signal that appropriately. As with x86 macOS, this can be reviewed later, but as we approach an initial alpha release, it's better to err on the side of not having command-line options which are/were largely debugging artifacts enshrined in the supported set.