aristo: switch to vector memtable
Every time we persist, we collect all changes into a batch and write
that batch to a memtable which rocksdb lazily will write to disk using a
background thread.

The default implementation of the memtable in rocksdb is a skip list
which can handle concurrent writes while still allowing lookups. We're
not using concurrent inserts and the skip list comes with significant
overhead both when writing and when reading.

Here, we switch to a vector memtable which is faster to write but
terrible to read. To compensate, we then proceed to flush the memtable
eagerly to disk which is a blocking operation.

One would think that the blocking of the main thread this would be bad
but it turns out that creating the skip list, also a blocking operation,
is even slower, resulting in a net win.

Coupled with this change, we also make the "lower" levels bigger
effectively reducing the average number of levels that must be looked at
to find recently written data. This could lead to some write
amplicification which is offset by making each file smaller and
therefore making compactions more targeted.

Taken together, this results in an overall import speed boost of about
3-4%, but above all, it reduces the main thread blocking time during
persist.