Revert "defer gc during block processing (#3384)" (#3445)
Deferred GC seemed like a good idea to reduce the amount of work done
during block processing, but a side effect of this is that more memory
ends up being allocated in certain workloads which in turn causes an
overall slowdown, with a long test showing a net performance effect that
hovers around 0% and more memory usage.

In particular, the troublesome range around 2M sees a 10-15% slowdown
and an ugly memory usage spike.

Reverting for now - it might be worth revisiting in the future under
different memory allocation patters, but as usual, it's better to not do
work at all (like in #3444) than to do work faster.

This reverts commit 3a009158d3e381562cd4ff855033533dbeadd404.