reuse VertexRef instance on update (#3444)
When updates to the MPT happen, a new VertexRef is allocated every time - this
keeps the code simple but has the significant downside that updates cause
unnecessary allocations.

Instead of allocating a new `VertexRef` on every update, we can update the
existing one provided that it is not shared. We can prevent it from being shared
by copying it eagerly when it's added to the layer. A downside of this approach
is that we also have to make a copy when invalidating hash keys, which affects
branch and account nodes mainly.

The tradeoff seems well worth it though, specially for imports that clock a nice
perf boost, like in this little test:

```
(21005462, 21008193]  14.46  15.50  2,479.35  2,656.98  9m26s  8m48s   7.16%   7.16%   -6.69%
(21013654, 21016385]  15.28  16.14  2,523.74  2,665.83  8m56s  8m27s   5.63%   5.63%   -5.33%
(21021846, 21024577]  15.52  17.66  2,539.25  2,889.61  8m47s  7m43s  13.80%  13.80%  -12.12%

blocks: 16384, baseline: 27m10s, contender: 24m59s
Time (total): -2m10s, -8.00%
```