fix slow startup during sync (#3514)
Unlike the rest of the database code, sync does not bulk-flush what it
writes and therefore runs into a pathologically slow case when a vector
memtable is used and it tries to read from it in a loop mingled with
writes.

This can cause the application to freeze for several seconds  - keep
vector for state database but otherwise use the skiplist.