Deprecate '--custom-network' and merge it's functionality to '--network'
We cannot remove `--custom-network` yet, there are still two users
depends on this: hive tests, and kurtosis interop test.

We an remove/disable `--custom-network` after both of hive tests and kurtosis interop
test accept the PR to fix them.

Now, the `--custom-network` only deprecated and not printed on CLI
anymore.