Fix WAL file buildup (#3603)
Due to the wrong CF being passed to the database flush mechanism during
sync, the sync KVT would remain unflushed as a memtable and thus prevent
the WAL from being released - given that a new wal is created for every
flush, this would lead to a buildup of small WAL files (50k and counting
on my hoodi node) after a sync.

At the same time, we don't really need to flush the KVT at every commit
since the skiplist permits efficient lookups (unlike the vector memtable
used for bulk updates to the state).