Add EL debug_getBlockHeader and use for Portal block verification
Portal no longer supports getting headers in its default history
network. That's fine as the ELs will keep storing the header chain
(for now) and thus does not need them. Portal however does need
access to the headers in order to be able to verify bodies and
receipts when those are being offered.

For this we had added a get header callback to the history
network. In this PR we fill in the used get header callback to
use the here newly added debug_getBlockHeader JSON-RPC method.

In case no web3 provider (EL) supporting this (only Nimbus) is
provided, the node can still run but will not be able to accept &
store any offered block bodies and receipts.