Fix invalid result error access + move empty txs check (#3649)
* Fix invalid result error access + move empty txs check

Check for empty txs in getBlockTransactionData and
getExecutionPayloadBodyV1.

First step as more locations potentially require this on usage of
also getBlockTransactions and getBlockTransactionHashes.

Can also add additionally check for rootHash calc.

* Fix era import by adjusting getUncleHashes to not use getBlockBody

When era based import does not store block bodies (transactions here
specifically), then this retrieval will fail when using getBlockBody.

Rework instead to just get the uncles and not care about the other
parts of the block body.

* Fix txRoot check bug as zeroHash32 instead of emptyRoot was used

zeroHash32 check was used originally and copied in this PR.
This is incorrect as the root of an empty tree is not the same.