Portal: Change JSON-RPC port defaults client/bridge (#3700)
Use to have port 8545/8546 for http/ws to be consistent with ELs
as Portal client was considered light client alternative for ELs
and this way devs don't have to adjust ports.

However, now the only use case is integration with an EL so it
becomes annoying to have colliding ports when you actually have to
run both together.