Portal json-rpc: Fix for having seqNum encoded as hexstring (#3710)
It seems that the web3/conversions import that got added
in commit 31ebce2076ee22b09bebc1acb195c5b814d64431 caused
JrpcConv leakage for uint64 making it encode as a hex string
(like Quantity) instead of the expected json number.

Working around this here with a new distinct uint64 (as we
cannot replace the JrpcConv for json-rpc server).

Also trying to add some testing for this on the portal common
rpc API similar as is done for discv5 + some cleanup on the latter