Portal + era1: Remove use of hardcoded mainnet mergeBlockNumber
Make it configurable in historical_hashes_accumulator and era1
code. This way it can be used for different networks (testnets).

As a result we also don't have to separately compile the
historical_hashes_accumulator tests with a different compile
time define.

Also tested nimbus_execution_client import for Sepolia in the
process and found an issue with the era1 files there which is
why an additional check in era1 open call got commented out for now.

Additional things that came up:
- Time to bring the two eraDBs together (Portal + EL).
- Time to bring Portal and EL ChainConfig together but the latter
could use some cleanup