NVP: Adjust CI to run also on EL changes
The nimbus_verified_proxy uses several modules from the
execution_chain folder. To make sure nothing breaks the CI should
also run on changes in that folder.

It uses nothing of the portal folder, thus CI can avoid to be run
on changes in portal.

The nimbus_execution_client doesn't use anthing from the
nimbus_verified_proxy folder, so the reverse relationship does
not need to exist.

This is different for portal, where EL does integrate that codebase.