Quickfix for connectingNodes leak on certain connection failures
On failures at postHelloSteps a connectingNodes leak will occur.
Not only will this just keep increasing connectingNodes, it also
means that these nodes will never be able to connect again or be
connected to again.

Quickfix as a cleaner fix might be putting the connectingNodes
incl/excl earlier or even outside of the rlpxAccept.